backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.6.3
pydantic==2.7.1
python-dotenv==1.0.1
dnspython==2.6.1